Clean up the code and move the ajax_xelon view in the PSA app
